Added sample.js that needs the csrf token.
This commit is only to make it easier to test the real change.
The sample.js file is automatically included, in the same way that protect.js is included.
It calls @@sample with a random value and sets this on the context.
That should kick in the auto csrf checking.
Note that you do always get a success message back.
But if it really fails, you see a Forbidden error in the log, and the existing number will stay the same: the transaction is aborted.
Note: I have seen this fail for the admin user, but that may be because this is a Zope level user.