Removed `RedirectTo` patch.
The patch has been merged to `Products.CMFFormController` 3.0.7 (Plone
4.3 and 5.0) and 3.1.2 (Plone 5.1).

Note that we are not requiring those versions in our `setup.py`,
because the code in this package no longer needs it.

We actually never had Products.CMFFormController in our
install_requires, so it would seem strange to begin with that now that
the code that we needed the dependency for has been removed. :-)